+++ CHANGELOG.md
@@ -31,10 +31,11 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - fix undefined `error.error` instances in various places
 - revert to using angular http `withInterceptorsFromDi`
 - fix type of ngbtooltip `openDelay` in angular templates
+- updated dependencies
 
 ### Homebridge Dependencies
 
-- `@homebridge/hap-client` @ `v3.0.0`
+- `@homebridge/hap-client` @ `v3.1.1`
 - `@homebridge/node-pty-prebuilt-multiarch` @ `v0.13.1`
 - `@homebridge/plugin-ui-utils` @ `v2.1.0`